Reorganise AST & AST builder types
