+++ src/parser/__tests__/declarations/functionDefinition.ts
@@ -1,5 +1,5 @@
 import { parse } from '../../parser';
-import { type Program } from '../../../ast/types';
+import { type Program } from '../../../ast/types/ast';
 
 describe('function definition', () => {
   test('handles function definitions without parameters', () => {
