+++ src/parser/__tests__/expressions/equalityExpression.ts
@@ -1,5 +1,5 @@
 import { parse } from '../../parser';
-import { type Program } from '../../../ast/types';
+import { type Program } from '../../../ast/types/ast';
 
 describe('equality expression', () => {
   test('handles equality', () => {
