+++ src/parser/__tests__/expressions/exclusiveOrExpression.ts
@@ -1,5 +1,5 @@
 import { parse } from '../../parser';
-import { type Program } from '../../../ast/types';
+import { type Program } from '../../../ast/types/ast';
 
 describe('exclusive or expression', () => {
   test('handles bitwise XOR', () => {
