+++ src/parser/__tests__/expressions/logicalOrExpression.ts
@@ -1,5 +1,5 @@
 import { parse } from '../../parser';
-import { type Program } from '../../../ast/types';
+import { type Program } from '../../../ast/types/ast';
 
 describe('logical or expression', () => {
   test('handles logical OR', () => {
