+++ src/parser/__tests__/expressions/postfixExpression.ts
@@ -1,5 +1,5 @@
 import { parse } from '../../parser';
-import { type Program } from '../../../ast/types';
+import { type Program } from '../../../ast/types/ast';
 
 describe('postfix expression', () => {
   test('handles primary expressions', () => {
