+++ src/parser/__tests__/expressions/primaryExpression.ts
@@ -1,5 +1,5 @@
 import { parse } from '../../parser';
-import { type Program } from '../../../ast/types';
+import { type Program } from '../../../ast/types/ast';
 
 describe('primary expression', () => {
   test('handles identifiers', () => {
